Extension: Exclude special pages and log to console
Make sure the extension itself is not loading in any special page.
The code is already there to make the button not appear, but it is
best if the browser knows not to load the extension files at all
for excluded namespaces. The biggest and most problematic namespace
is the `Special:` namespace, that includes within it pages like
`Special:Login` and `Special:Preferences` where we don't want the
code to load at all.

In the case of a gadget, the MediaWiki system knows not to load
the gadget on "sensitive" pages (like `Special:Preferences` and
`Special:Login`) so those, at least, are covered, and the code
itself verifies that the button is not injected into the DOM in
pages that are not articles.