Only load welcome popup on valid pages
While the extension only actually loads on valid pages, there are
cases (like talk page and history, etc) where the files themselves
are loaded but the script won't when the activation singleton
checks for valid pages.

However, that check is only done for initialization of the main
script, and not for the welcome popup.

More importantly, the gadget always loads the welcome popup, and
we have to make sure it does not run or display it for invalid
pages since the validity check was only done for the activation
initialization of the general script.

This commit adds that check to the welcome popup as well, so it
does not load if the page is not valid.

Bonus: Have the 'isValidPage' method only check if the page is
valid, rather than also check if the system was also initialized,
so that we can repeatedly make this check even after initialization.