WelcomePopup: Fix positioning problem for long pages
When we have long pages where the "tools" menu is under the viewport,
the `<html>` tag, unhelpfully, stops (its height is limited) at the
edge of the viewport. This means that the `<body>` is similarly limited
in height, even though the content itself keeps going.

As a result, the popup, when instantiated, cannot be placed below
that point, which creates a popup that appears above where it should.

This is a hacky way to fix this. When and if the popup is active,
the code adds a class to the `html` block and sets it at a large
min-height that will cover at least the side menu to be able to
get the popup to appear properly.

HTML sucks.