Fix highlighting cutting off font kerning
Add a ::before pseudo element to the 'active' CSS class, which is used to highlight when hovering
or selecting an element. When the .active class is set, a pseudo element is added behind it with
a background color that highlights the text. This behind-element approach will prevent any content
from being obscured.

Bug: T232490