Cleanup: Drop jquery fallback, only use in tests
The initial need for `const $ = ...` was for making sure that
tests have that package, while not polluting the general scripts
with an unnecessary overload of jquery that already exists in
mediawiki.

Instead of doing that, we can use the testHelper's mocking of the
globals to pretend `window.$` is set to jquery, which would mean
it's already available all the time.

Props to @samwilson on questioning the original approach.