Separate Firefox and Chrome manifests
This adds the Gecko ID back into manifest.json, and creates a
new build process that uses separate manifest files for Chrome
and Firefox, outputting two zip files for upload to the web
stores.

We removed the Gecko ID because I thought it wasn't required and
it was causing a warning in Chrome. Turns out the storage.sync API
isn't available unless you set the ID.

https://phabricator.wikimedia.org/T236226
https://phabricator.wikimedia.org/T232460

Bug: T236226
Bug: T232460