Gray out non-tokenized elements
Non-tokenized elements have no data associated with them, so we gray
them out to give visual indication of this. Hovering over the grayed-out
elements will increase the opacity to 80%.

There doesn't appear to be a rock-solid solution to identifying any and
all untokenized elements. So we loop through the DOM and pick out ones
that don't have .editor-token as a child. Additionally we have a
manually maintained list of selectors under `untokenizedElements` in
config.js, which covers some elements that are nested deeper than the
first level. This is not comprehensive but can be updated as needed.

Bug: T235130