Prevent content script from being injected more than once
Although I haven't quite figured out what's causing it, there are
times when the content script is injected twice. This adds a
guard against this, and even if it's not tackling the root cause
it shoudln't be too harmful a check to add.

This also removes the 'content_scripts.all_frames' setting from
manifest.json, returning it to its default value of false (meaning
that the script will only be loaded for the topmost frame in a
tab). This shouldn't change anything, because Wikipedia never
loads itself in a frame, but it seemed like a good guess as to
why things were loading twice (they still were even without it...).

https://phabricator.wikimedia.org/T232460

Bug: T232460