Move click handler from button to the link inside the button
This might seem like an overly-large change for this little tweak,
but most of what it's doing is making the nomenclature for the
button and the link more consistent. The button is the `<li>` and
the link is the `<a>` inside it.

This patch also moves the creation of the portlet into the
getButton(), so we don't have the odd create-then-retrieve flow
in initialize(). And it turns both button and link into jQuery
objects, so we don't have to think too hard about whether they're
jQuery or HTMLElements (could have gone the other way, and done
them both as HTMLElements, but we want to use jQuery's toggle()
on them, so this is less code).

https://phabricator.wikimedia.org/T239842

Bug: T239842