Refactor RevisionPopupWidget for MVC
We've changed the system to use MVC but didn't really touch the
widget, so it had a lot of its own logic. That logic didn't quite
fit the new features we want (and are about to tweak) like extending
some of the pending animation to more than the comment, or adding
more of the dynamic information within the promises and events.

This commit refactors the widget to listen to the model and respond
to refresh events, and have the click event trigger the controller
to utilize the API. It means, among other things, that --
- We don't need to deliver extra data like isCached between the
  MVC barriers,
- It should be a little more organized with some clear order of
  operations
- The popup doesn't constantly get rebuilt -- it builds its structure
  and only replaces the actual data when updated.

Bug: https://phabricator.wikimedia.org/T241004